expand {reshape}	R Documentation

Expand

Description

Expand out condensed data frame.

Usage

expand(data)

Arguments

data
condensed data frame

Details

If aggregating function supplied to condense returns multiple values, this function "melts" it again, creating a new column called result\_variable.

If the aggregating funtion is a named vector, then those names will be used, otherwise will be number X1, X2, ..., Xn etc.

Author(s)

Hadley Wickham <h.wickham@gmail.com>

[Package reshape version 0.8.8 Index]