funstofun {reshape}	R Documentation

Aggregate multiple functions into a single function

Description

Combine multiple functions to a single function returning a named vector of outputs

Usage

funstofun(...)

Arguments

...
functions to combine

Details

Each function should produce a single number as output

Author(s)

Hadley Wickham <h.wickham@gmail.com>

Examples

funstofun(min, max)(1:10)
funstofun(length, mean, var)(rnorm(100))

[Package reshape version 0.8.8 Index]