melt {reshape}	R Documentation

Melt

Description

Melt an object into a form suitable for easy casting.

Usage

melt(data, ...)

Arguments

data
Data set to melt

...
Other arguments passed to the specific melt method

Details

This the generic melt function. See the following functions for specific details for different data structures:

melt.data.frame for data.frames

melt.array for arrays, matrices and tables

melt.list for lists

Author(s)

Hadley Wickham <h.wickham@gmail.com>

[Package reshape version 0.8.10 Index]