melt.cast_df {reshape}	R Documentation

Melt cast data.frames

Description

Melt the results of a cast

Usage

## S3 method for class 'cast_df'
melt(data, drop.margins=TRUE, ...)

Arguments

data

drop.margins

...
other arguments ignored

Details

This can be useful when performning complex aggregations - melting the result of a cast will do it's best to figure out the correct variables to use as id and measured.

Author(s)

Hadley Wickham <h.wickham@gmail.com>

[Package reshape version 0.8.8 Index]