namerows {reshape}	R Documentation

Name rows

Description

Add variable to data frame containing rownames

Usage

namerows(df, col.name = "id")

Arguments

df
data frame

col.name
name of new column containing rownames

Details

This is useful when the thing that you want to melt by is the rownames of the data frame, not an explicit variable

Author(s)

Hadley Wickham <h.wickham@gmail.com>

[Package reshape version 0.8.8 Index]