prettyprint {reshape}	R Documentation

Pretty print

Description

Print reshaped data frame

Usage

prettyprint(x, digits=getOption("digits"), ..., colnames=TRUE)

Arguments

x

digits

...

colnames

Details

This will always work on the direct output from cast, but may not if you have manipulated (e.g. subsetted) the results.

Author(s)

Hadley Wickham <h.wickham@gmail.com>

[Package reshape version 0.8.8 Index]